Moved AssetManagerUtil back to utilities.
