+++ src/main/java/roborally/listeners/CollisionListener.java
@@ -3,7 +3,7 @@ package roborally.listeners;
 import com.badlogic.gdx.math.GridPoint2;
 import roborally.game.robot.Robot;
 import roborally.gameview.layout.ILayers;
-import roborally.utilities.assets.AssetManagerUtil;
+import roborally.utilities.AssetManagerUtil;
 import roborally.utilities.SettingsUtil;
 import roborally.utilities.enums.LayerName;
